chore(nexus): nexus code clean up
