+++ io-engine/src/bdev/nexus/nexus_bdev.rs
@@ -12,24 +12,23 @@ use std::{
 };
 
 use futures::channel::oneshot;
-use nix::errno::Errno;
 use serde::Serialize;
-use snafu::{ResultExt, Snafu};
-use tonic::{Code, Status};
+use snafu::ResultExt;
 use uuid::Uuid;
 
 use super::{
+    nexus_err,
     nexus_lookup_name_uuid,
     nexus_submit_request,
-    ChildError,
     ChildState,
     DrEvent,
+    Error,
     NbdDisk,
-    NbdError,
     NexusChannel,
     NexusChild,
     NexusModule,
     PersistOp,
+    VerboseError,
 };
 
 use crate::{
@@ -37,7 +36,6 @@ use crate::{
         device_destroy,
         nexus::{nexus_persistence::PersistentNexusInfo, NexusIoSubsystem},
     },
-    bdev_api::BdevError,
     core::{
         Bdev,
         BdevHandle,
@@ -50,8 +48,7 @@ use crate::{
         Share,
         MWQ,
     },
-    rebuild::RebuildError,
-    subsys::{NvmfError, NvmfSubsystem},
+    subsys::NvmfSubsystem,
 };
 
 use spdk_rs::{
@@ -67,296 +64,16 @@ use spdk_rs::{
 pub static NVME_MIN_CNTLID: u16 = 1;
 pub static NVME_MAX_CNTLID: u16 = 0xffef;
 
-/// Obtain the full error chain
-pub trait VerboseError {
-    fn verbose(&self) -> String;
-}
-
-impl<T> VerboseError for T
-where
-    T: std::error::Error,
-{
-    /// loops through the error chain and formats into a single string
-    /// containing all the lower level errors
-    fn verbose(&self) -> String {
-        let mut msg = format!("{}", self);
-        let mut opt_source = self.source();
-        while let Some(source) = opt_source {
-            msg = format!("{}: {}", msg, source);
-            opt_source = source.source();
-        }
-        msg
-    }
-}
-
-/// Common errors for nexus basic operations and child operations
-/// which are part of nexus object.
-#[derive(Debug, Snafu)]
-#[snafu(visibility(pub(crate)), context(suffix(false)))]
-pub enum Error {
-    #[snafu(display("Nexus {} does not exist", name))]
-    NexusNotFound { name: String },
-    #[snafu(display("Nexus {} exists and is initialising", name))]
-    NexusInitialising { name: String },
-    #[snafu(display("Invalid nexus uuid \"{}\"", uuid))]
-    InvalidUuid { uuid: String },
-    #[snafu(display(
-        "Nexus uuid \"{}\" already exists for nexus \"{}\"",
-        uuid,
-        nexus
-    ))]
-    UuidExists { uuid: String, nexus: String },
-    #[snafu(display("Nexus with name \"{}\" already exists", name))]
-    NameExists { name: String },
-    #[snafu(display("Invalid encryption key"))]
-    InvalidKey {},
-    #[snafu(display("Failed to create crypto bdev for nexus {}", name))]
-    CreateCryptoBdev { source: Errno, name: String },
-    #[snafu(display("Failed to destroy crypto bdev for nexus {}", name))]
-    DestroyCryptoBdev { source: Errno, name: String },
-    #[snafu(display(
-        "The nexus {} has been already shared with a different protocol",
-        name
-    ))]
-    AlreadyShared { name: String },
-    #[snafu(display("The nexus {} has not been shared", name))]
-    NotShared { name: String },
-    #[snafu(display("The nexus {} has not been shared over NVMf", name))]
-    NotSharedNvmf { name: String },
-    #[snafu(display("Failed to share nexus over NBD {}", name))]
-    ShareNbdNexus { source: NbdError, name: String },
-    #[snafu(display("Failed to share nvmf nexus {}", name))]
-    ShareNvmfNexus { source: CoreError, name: String },
-    #[snafu(display("Failed to unshare nexus {}", name))]
-    UnshareNexus { source: CoreError, name: String },
-    #[snafu(display(
-        "Failed to register IO device nexus {}: {}",
-        name,
-        source
-    ))]
-    RegisterNexus { source: Errno, name: String },
-    #[snafu(display("Failed to create child of nexus {}: {}", name, source))]
-    CreateChild { source: BdevError, name: String },
-    #[snafu(display("Deferring open because nexus {} is incomplete", name))]
-    NexusIncomplete { name: String },
-    #[snafu(display(
-        "Child {} of nexus {} is too small: size = {} x {}",
-        child,
-        name,
-        num_blocks,
-        block_size
-    ))]
-    ChildTooSmall {
-        child: String,
-        name: String,
-        num_blocks: u64,
-        block_size: u64,
-    },
-    #[snafu(display("Children of nexus {} have mixed block sizes", name))]
-    MixedBlockSizes { name: String },
-    #[snafu(display(
-        "Child {} of nexus {} has incompatible size or block size",
-        child,
-        name
-    ))]
-    ChildGeometry { child: String, name: String },
-    #[snafu(display("Child {} of nexus {} cannot be found", child, name))]
-    ChildMissing { child: String, name: String },
-    #[snafu(display("Child {} of nexus {} has no error store", child, name))]
-    ChildMissingErrStore { child: String, name: String },
-    #[snafu(display(
-        "Failed to acquire write exclusive reservation on child {} of nexus {}",
-        child,
-        name
-    ))]
-    ChildWriteExclusiveResvFailed {
-        source: ChildError,
-        child: String,
-        name: String,
-    },
-    #[snafu(display("Failed to open child {} of nexus {}", child, name))]
-    OpenChild {
-        source: ChildError,
-        child: String,
-        name: String,
-    },
-    #[snafu(display("Failed to close child {} of nexus {}", child, name))]
-    CloseChild {
-        source: BdevError,
-        child: String,
-        name: String,
-    },
-    #[snafu(display(
-        "Cannot delete the last child {} of nexus {}",
-        child,
-        name
-    ))]
-    DestroyLastChild { child: String, name: String },
-    #[snafu(display(
-        "Cannot remove the last child {} of nexus {} from the IO path",
-        child,
-        name
-    ))]
-    DestroyLastHealthyChild { child: String, name: String },
-    #[snafu(display(
-        "Cannot remove the last healthy child {} of nexus {} from the IO path",
-        child,
-        name
-    ))]
-    RemoveLastChild { child: String, name: String },
-    #[snafu(display(
-        "Cannot fault the last healthy child {} of nexus {}",
-        child,
-        name
-    ))]
-    FaultingLastHealthyChild { child: String, name: String },
-    #[snafu(display("Failed to destroy child {} of nexus {}", child, name))]
-    DestroyChild {
-        source: BdevError,
-        child: String,
-        name: String,
-    },
-    #[snafu(display("Child {} of nexus {} not found", child, name))]
-    ChildNotFound { child: String, name: String },
-    #[snafu(display("Child {} of nexus {} already exists", child, name))]
-    ChildAlreadyExists { child: String, name: String },
-    #[snafu(display("Failed to pause child {} of nexus {}", child, name))]
-    PauseChild { child: String, name: String },
-    #[snafu(display("Suitable rebuild source for nexus {} not found", name))]
-    NoRebuildSource { name: String },
-    #[snafu(display(
-        "Failed to create rebuild job for child {} of nexus {}",
-        child,
-        name,
-    ))]
-    CreateRebuild {
-        source: RebuildError,
-        child: String,
-        name: String,
-    },
-    #[snafu(display(
-        "Rebuild job not found for child {} of nexus {}",
-        child,
-        name,
-    ))]
-    RebuildJobNotFound { child: String, name: String },
-    #[snafu(display(
-        "Rebuild job already exists for child {} of nexus {}",
-        child,
-        name,
-    ))]
-    RebuildJobAlreadyExists { child: String, name: String },
-    #[snafu(display(
-        "Failed to execute rebuild operation on job {} of nexus {}",
-        job,
-        name,
-    ))]
-    RebuildOperation {
-        job: String,
-        name: String,
-        source: RebuildError,
-    },
-    #[snafu(display("Invalid ShareProtocol value {}", sp_value))]
-    InvalidShareProtocol { sp_value: i32 },
-    #[snafu(display("Invalid NvmeAnaState value {}", ana_value))]
-    InvalidNvmeAnaState { ana_value: i32 },
-    #[snafu(display("Invalid arguments for nexus {}: {}", name, args))]
-    InvalidArguments { name: String, args: String },
-    #[snafu(display("Failed to create nexus {}", name))]
-    NexusCreate { name: String },
-    #[snafu(display("Failed to destroy nexus {}", name))]
-    NexusDestroy { name: String },
-    #[snafu(display(
-        "Child {} of nexus {} is not degraded but {}",
-        child,
-        name,
-        state
-    ))]
-    ChildNotDegraded {
-        child: String,
-        name: String,
-        state: String,
-    },
-    #[snafu(display("Failed to get BdevHandle for snapshot operation"))]
-    FailedGetHandle,
-    #[snafu(display("Failed to create snapshot on nexus {}", name))]
-    FailedCreateSnapshot { name: String, source: CoreError },
-    #[snafu(display("NVMf subsystem error: {}", e))]
-    SubsysNvmf { e: String },
-    #[snafu(display("failed to pause {} current state {:?}", name, state))]
-    Pause {
-        state: NexusPauseState,
-        name: String,
-    },
-}
-
-impl From<NvmfError> for Error {
-    fn from(error: NvmfError) -> Self {
-        Error::SubsysNvmf {
-            e: error.to_string(),
-        }
-    }
-}
-
-impl From<Error> for tonic::Status {
-    fn from(e: Error) -> Self {
-        match e {
-            Error::NexusNotFound {
-                ..
-            } => Status::not_found(e.to_string()),
-            Error::InvalidUuid {
-                ..
-            } => Status::invalid_argument(e.to_string()),
-            Error::InvalidKey {
-                ..
-            } => Status::invalid_argument(e.to_string()),
-            Error::AlreadyShared {
-                ..
-            } => Status::invalid_argument(e.to_string()),
-            Error::NotShared {
-                ..
-            } => Status::invalid_argument(e.to_string()),
-            Error::NotSharedNvmf {
-                ..
-            } => Status::invalid_argument(e.to_string()),
-            Error::CreateChild {
-                ..
-            } => Status::invalid_argument(e.to_string()),
-            Error::MixedBlockSizes {
-                ..
-            } => Status::invalid_argument(e.to_string()),
-            Error::ChildGeometry {
-                ..
-            } => Status::invalid_argument(e.to_string()),
-            Error::OpenChild {
-                ..
-            } => Status::invalid_argument(e.to_string()),
-            Error::DestroyLastChild {
-                ..
-            } => Status::invalid_argument(e.to_string()),
-            Error::ChildNotFound {
-                ..
-            } => Status::not_found(e.to_string()),
-            e => Status::new(Code::Internal, e.to_string()),
-        }
-    }
-}
-
 pub(crate) static NEXUS_PRODUCT_ID: &str = "Nexus CAS Driver v0.0.1";
 
+/// TODO
 #[derive(Debug)]
 pub enum NexusTarget {
     NbdDisk(NbdDisk),
     NexusNvmfTarget,
 }
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum NexusPauseState {
-    Unpaused,
-    Pausing,
-    Paused,
-    Unpausing,
-}
 
+/// TODO
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub enum NvmeAnaState {
     InvalidState, // invalid, do not use
@@ -463,6 +180,7 @@ pub struct Nexus<'n> {
     _pin: PhantomPinned,
 }
 
+/// Nexus status enumeration.
 #[derive(Debug, Serialize, Clone, Copy, PartialEq, PartialOrd)]
 pub enum NexusStatus {
     /// The nexus cannot perform any IO operation
@@ -473,6 +191,7 @@ pub enum NexusStatus {
     Online,
 }
 
+/// Nexus state enumeration.
 #[derive(Debug, Serialize, Clone, Copy, PartialEq, PartialOrd)]
 pub enum NexusState {
     /// nexus created but no children attached
@@ -818,7 +537,7 @@ impl<'n> Nexus<'n> {
                     }
                 }
                 nex.as_mut().set_state(NexusState::Closed);
-                Err(err).context(RegisterNexus {
+                Err(err).context(nexus_err::RegisterNexus {
                     name: nex.name.clone(),
                 })
             }
@@ -1302,7 +1021,7 @@ impl<'n> BdevOps for Nexus<'n> {
     fn dump_info_json(&self, w: JsonWriteContext) {
         w.write_named_array_begin("children");
         if let Err(err) = w.write(&self.children) {
-            error!("Failed to dump into JSON: {}", err);
+            error!("Failed to dump into JSON: {}", err.to_string());
         }
         w.write_array_end();
     }
@@ -1450,7 +1169,9 @@ async fn nexus_create_internal(
         {
             error!(
                 "failed to create nexus {}: failed to create child {}: {}",
-                name, child, error
+                name,
+                child,
+                error.to_string()
             );
             nexus_bdev.data_mut().close_children().await;
 
@@ -1486,11 +1207,10 @@ async fn nexus_create_internal(
         }
 
         Err(error) => {
-            error!("failed to open nexus {}: {}", name, error);
+            error!("failed to open nexus {}: {}", name, error.to_string());
             nexus_bdev.data_mut().close_children().await;
             Err(error)
         }
-
         Ok(_) => Ok(()),
     }
 }
